polyadic_preprocessing/requirements.txt
annotated-types==0.5.0
annoy==1.17.1
appdirs==1.4.4
apturl==0.5.2
argcomplete==3.1.2
asgiref==3.7.2
attrs==21.2.0
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
black==23.9.1
blinker==1.6.2
blosc2==2.0.0
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.2.0
ciso8601==2.3.1
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
-e git+https://github.com/commonsense/conceptnet5@fda1b39607ea6f07e41ddce22d9d5f8360ded581#egg=ConceptNet&subdirectory=../../../submodule/conceptnet5
ConfigArgParse==1.5.3
cryptography==41.0.2
cssselect==1.2.0
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.33
datamodel-code-generator==0.21.5
datrie==0.8.2
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
deprecation==2.1.0
devscripts===2.22.1ubuntu1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
dot2tex==2.11.3
duplicity==0.8.21
email-validator==2.0.0.post2
fasteners==0.14.1
Flask==2.0.0
flloat==0.3.0
fonttools==4.29.1
fs==2.4.12
ftfy==6.1.1
future==0.18.2
gast==0.5.2
genson==1.2.2
ghp-import==2.1.0
git-filter-repo==2.34.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.8.0
google-auth-oauthlib==0.5.2
gpg==1.16.0
graphml2svg==0.0.4
graphviz==0.20
greenlet==2.0.1
grip==4.6.1
gyp==0.1
h5py==3.8.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
imageio==2.19.1
img2pdf==0.4.4
importlib-metadata==4.6.4
imutils==0.5.4
inflect==5.6.2
ipadic==1.0.0
isort==5.12.0
itsdangerous==2.1.2
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.0.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-spec==0.1.6
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.2
langcodes==3.3.0
language-selector==0.1
lark==1.1.7
lark-parser==0.12.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lazy_loader==0.3
livereload==2.6.3
lockfile==0.12.2
logaut @ file:///home/giacomo/projects/reducer/submodules/logaut
louis==3.20.0
ltlf2dfa @ git+https://github.com/whitemech/LTLf2DFA.git@08bde1ea75ff513ddd7fc3848f5499db7a60418f
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==0.7.8
Markdown==3.4.3
markdown2==2.4.10
MarkupSafe==2.1.2
MaterialX==0.0.0.dev1
matplotlib==3.5.2
mecab-python3==1.0.6
meld==3.20.4
mercurial==6.1.1
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-material-extensions==1.1.1
monotonic==1.6
more-itertools==8.10.0
Morfessor==2.0.6
mpmath==1.3.0
msgpack==1.0.4
msgpack-python==0.5.6
mutagen==1.46.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.1
notify2==0.3
numexpr==2.8.4
numpy==1.26.0
oauthlib==3.2.0
ocrmypdf==14.3.0
olefile==0.46
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.7
opencv-python==4.8.0.74
opyenxes==0.3.0
ordered-set==4.1.0
osfclient==0.0.5
packaging==23.1
paginate==0.5.6
pandas==1.4.2
pandasql==0.7.3
paramiko==2.9.3
path-and-address==2.0.1
pathable==0.4.3
pathspec==0.11.2
patsy==0.5.3
pdfminer.six==20221105
pexpect==4.8.0
pikepdf==8.2.1
Pillow==9.0.1
platformdirs==3.10.0
plotille==5.0.0
plotly==5.13.0
pluggy==1.2.0
ply==3.11
prance==23.6.21.0
progressbar==2.5
progressbar2==4.0.0
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pyconll==3.2.0
pycparser==2.21
pycryptodomex==3.18.0
pycups==2.0.1
pydantic==2.3.0
pydantic_core==2.6.3
pydyf==0.1.2
Pygments==2.16.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pylogics==0.2.1
pymacaroons==0.13.0
pymdown-extensions==10.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyphen==0.12.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyquery==2.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
PySnooper==1.2.0
pysrt==1.1.2
pythomata==0.3.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
python-utils==3.2.2
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyyaml_env_tag==0.1
ratelimiter==1.2.0.post0
readtime==3.0.0
regex==2022.10.31
reindent==3.5.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rsa==4.8
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.2.1
scipy==1.11.2
scour==0.38.2
scp==0.14.5
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
Shapely==1.8.2
six==1.16.0
smmap==5.0.0
snakemake==5.5.4
soupsieve==2.3.1
SQLAlchemy==1.4.45
sqlparse==0.4.4
srt==3.5.3
statsmodels==0.13.5
SumoNetVis==1.6.0
svgwrite==1.4.3
sympy==1.12
systemd-python==234
tables==3.8.0
tenacity==8.2.1
tesserocr==2.6.1
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2023.7.18
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unidiff==0.5.5
urllib3==1.26.5
usb-creator==0.3.7
vosk==0.3.45
wadllib==1.3.6
warlock==2.0.1
watchdog==3.0.0
wcwidth==0.2.6
weasyprint==54.1
webencodings==0.5.1
websocket-client==1.2.3
websockets==11.0.3
Werkzeug==2.3.4
wordfreq==3.0.3
wrapt==1.14.1
wxPython==4.1.1
xdg==5
xkit==0.0.0
xmltodict==0.11.0
yt-dlp==2023.7.6
zipp==1.0.0
